// app.client/src/api/models/UserListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserListVM = {
  userID?: string | null;
  email?: string | null;
  userName?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  isActive?: boolean;
  createdDate?: string;
  id?: string | null;
  reportToUser?: string | null;
};